fix: move keys to api-keys for consistency and to fix broken link
