+++ src/routes/(console)/project-[region]-[project]/overview/api-keys/[key]/+page.ts
