+++ src/routes/(console)/project-[region]-[project]/overview/api-keys/[key]/+page@project-[region]-[project].svelte
