+++ src/routes/(console)/project-[region]-[project]/overview/api-keys/checkbox.svelte
